tests/factories.py
import factory

from students_manager.students.models import Student
from students_manager import db


class StudentFactory(factory.alchemy.SQLAlchemyModelFactory):
    class Meta:
        model = Student
        sqlalchemy_session = db.session

    id = factory.Faker("uuid4")
    name = factory.Faker("first_name")
    surname = factory.Faker("last_name")
    specialization = factory.Faker("word")
